requirements.txt
paramiko
prettytable
python-dotenv